requirements.txt
xmltodict==0.12.0
torch==1.6.0
PyYAML==5.3.1
tensorboard==2.3.0
flake8==3.8.3
pytest==6.0.1
fire==0.3.1
pandas==1.1.1
Flask==1.1.2
tornado==6.0.4
jupyter==1.0.0
jupyterlab==2.2.8
matplotlib==3.3.2
matplotlib-venn==0.11.5
tqdm==4.50.2